[DOM-363] Add user friendly errors descriptions
